progress in parametrization, error handling
