+++ mage_ai/cenabast/pipelines/product_data_pipeline/metadata.yaml
@@ -10,12 +10,12 @@ blocks:
   language: python
   name: pull_product_data
   retry_config: null
-  status: updated
+  status: executed
   timeout: null
   type: data_loader
   upstream_blocks: []
   uuid: pull_product_data
-- all_upstream_blocks_executed: false
+- all_upstream_blocks_executed: true
   color: null
   configuration: {}
   downstream_blocks:
@@ -26,7 +26,7 @@ blocks:
   language: python
   name: filter_product_data
   retry_config: null
-  status: executed
+  status: updated
   timeout: null
   type: transformer
   upstream_blocks:
@@ -42,7 +42,7 @@ blocks:
   language: python
   name: export_product_data
   retry_config: null
-  status: executed
+  status: updated
   timeout: null
   type: data_exporter
   upstream_blocks:
