+++ mage_ai/cenabast/utils/data_exporter/vendor_functions.py
@@ -20,6 +20,7 @@ def create_or_update_vendor(contract, api_clients, general_data):
     filtered_attributes = {key: value for key, value in existing_vendor.items() if key in payload.keys()}
     if payload != filtered_attributes:
       general_data['logger'].info(f"Vendor {vendor_run} needs update, updating.")
+      general_data['logger'].debug(f"Vendor payload to use: {payload}")
       api_response = api_clients['vendors_client'].update_vendor(existing_vendor['id'], payload)
       vendor = api_response.get('results', {})
     else:
@@ -28,6 +29,7 @@ def create_or_update_vendor(contract, api_clients, general_data):
   else:
     # Create vendor
     general_data['logger'].info(f"Vendor {vendor_run} didnt exist, creating.")
+    general_data['logger'].debug(f"Vendor payload to use: {payload}")
     api_response = api_clients['vendors_client'].create_vendor(payload)
     vendor = api_response.get('results', {})
     general_data['logger'].info(f"Created new vendor {vendor_run}")
